Initial change for onnxruntime-utils
