+++ js/package.json
@@ -27,7 +27,7 @@
   "scripts": {
     "prepare": "tsc --build scripts",
     "lint": "eslint . --ext .ts --ext .tsx",
-    "format:ts": "clang-format --glob=\"{scripts/**/*.ts,common/{lib,test}/**/*.ts,node/{lib,script,test}/**/*.ts,web/{lib,script,test}/**/*.ts,react_native/{android,example,ios,lib}/**/*.{ts,tsx}}\" --style=file -i",
+    "format:ts": "clang-format --glob=\"{scripts/**/*.ts,common/{lib,test}/**/*.ts,utils/{lib,test}/**/*.ts,node/{lib,script,test}/**/*.ts,web/{lib,script,test}/**/*.ts,react_native/{android,example,ios,lib}/**/*.{ts,tsx}}\" --style=file -i",
     "format:js": "clang-format --glob=\"{{,common,node,web,react_native}/{*,.*}.{,m,c}js,web/test/e2e/**/*.{,m,c}js}\" --style=file -i",
     "format:cf": "clang-format --glob=\"{node/src/**/*.{cc,h},react_native/{android,example,ios,lib}/**/*.{mm,java}}\" --style=file -i",
     "format:json": "prettier \"**/*.{json,jsonc}\" --write",
