+++ js/utils/README.md
@@ -1,0 +1,10 @@
+# ONNX Runtime Utils
+
+This library provides utilities for ONNX Runtime, including:
+- TBD
+- TBD
+- TBD
+
+## License
+
+License information can be found [here](https://github.com/microsoft/onnxruntime/blob/main/README.md#license).
